Fix Codeclimate review
