+++ src/polyswarmclient/ethereum/__init__.py
@@ -1,5 +1,5 @@
 from .balanceclient import BalanceClient
-from .bountiesclient import BountiesClient
+from .bountiesclient.base import BountiesClient
 from .offersclient import OffersClient
 from .relayclient import RelayClient
 from .stakingclient import StakingClient
