+++ src/polyswarmclient/fast/transaction.py
@@ -2,7 +2,6 @@ import logging
 from abc import ABCMeta, abstractmethod
 from typing import Any, Dict
 
-import aiohttp
 from polyswarmtransaction import SignedTransaction, Transaction
 from polyswarmclient import Client
 from polyswarmclient.exceptions import TransactionError
@@ -51,7 +50,7 @@ class PolySwarmTransactionRequest(metaclass=ABCMeta):
     async def post_transaction(self, signed: SignedTransaction, api_key: str) -> Dict[str, Any]:
         success, results = await self.client.make_request('POST', self.path, json=signed.payload, api_key=api_key,
                                                           chain='side')
-        # FIXME: Remove this temporary solution once there is time to use raise_for_status in make_request
+        # TODO: Remove this temporary solution once there is time to use raise_for_status in make_request
         if not success:
             logger.error('Error posting transaction: %s', results)
             raise TransactionError('Failed to post transaction')
